test(ws): make the Responses WebSocket end-to-end suite actually run in CI
`responses_websocket_e2e` was a `fn main()` scenario binary. CI ran
`cargo test -p aether-integration-tests --bins`, which compiled it and
ran zero tests, and nothing else invoked it — no workflow, script, or
doc referenced the binary or the JSON report it printed. The feature's
only end-to-end coverage was, in practice, dead code.

Move it to `tests/` as three `#[tokio::test]`s and add `--tests` to the
job's existing command, so the suite is executed rather than merely
compiled. The three tests run as parallel threads of one process under
`cargo test`; each builds its own mock upstream, SQLite file on a fresh
temp path, and gateway on a random port, so they do not interfere.

Two defects in the old harness surfaced immediately:

Usage recording was switched off. `AppState` defaults `usage_runtime` to
`UsageRuntime::disabled()`, and the harness never called
`with_usage_runtime_config`, so `record_stream_terminal` returned early
and no billing row was ever written. The scenario could not have caught
a lost usage row, which is precisely the class of bug the previous two
commits fixed. The harness now enables the runtime with queueing off, so
terminal writes land in-process rather than through Redis.

Waiting on "no longer pending" is not waiting for a settled row. A row
is created `Pending` at turn start and reaches its final shape through
several independent writes; in between it can read as completed with
zero tokens and no WebSocket metadata, because the stream-started
lifecycle event is dispatched without being awaited and is reduced to
the fixed allow-list in `retain_first_byte_request_metadata`. Against a
mock upstream that answers instantly, a naive wait returns that
intermediate row. Each test now waits for the end state it asserts.

The mock upstream is programmable so a turn can be interrupted, which
covers the two paths that leave no trace in the protocol:

- `continuation_reuses_one_upstream_connection_and_bills_both_turns`
  pins the headline guarantee — one physical upstream socket, the
  continuation id surviving provider body normalization — and now also
  asserts both turns persist independent, correctly-tokened rows
  flagged `is_websocket`.
- `client_disconnect_mid_turn_still_settles_the_usage_row` covers a
  client walking away mid-turn. The relay loop owns the turn and the
  upstream never reports this, so it is the path the `Drop` safety net
  in the previous commit exists for.
- `upstream_drop_mid_turn_reports_an_error_and_settles_the_usage_row`
  covers an upstream dying mid-turn: the client must be told, and the
  turn must still settle.

The assertions read from the database on purpose. A turn settles its
billing row from a task that outlives the relay loop, so a client that
saw `response.completed` is not evidence the turn was ever accounted
for — only the row is.

Suite runs in ~1.7s; 10 consecutive runs under the CI command were clean.
