+++ .github/workflows/rust-ci.yml
@@ -387,11 +387,11 @@ jobs:
       - name: Setup sccache
         uses: mozilla-actions/sccache-action@v0.0.9
 
-      - name: Test scenario binaries
+      - name: Test scenario binaries and end-to-end suites
         env:
           RUSTC_WRAPPER: sccache
           SCCACHE_GHA_ENABLED: "true"
-        run: cargo test -p aether-integration-tests --bins
+        run: cargo test -p aether-integration-tests --bins --tests
 
       - name: Show sccache stats
         if: always()
